python/requirements.txt
# MetaLens — Python sidecar dependencies
# Python 3.11–3.13

# Web server
fastapi==0.115.12
uvicorn==0.34.2

# Images
Pillow==11.2.1
piexif==1.1.3

# Audio / Video
mutagen==1.47.0

# PDF
pypdf==5.4.0

# Office Open XML
python-docx==1.2.0
openpyxl==3.1.5
python-pptx==1.0.2

# Legacy OLE (DOC/XLS/PPT)
olefile==0.47

# Generic fallback (pure Python, many formats)
hachoir==3.3.0

# Extended attributes — Linux/macOS only
xattr==1.1.0; sys_platform != "win32"

# Windows file attributes
pywin32==310; sys_platform == "win32"

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
